fix image content (not viewable)
Fixed the image content problem where if the images were not fit to the size of devices and were not viewable had to scroll down and up to view the whole big image. With this fix, The whole image is fitted perfectly. (also for different kinds of devices, now its easily viewable and looks good as image viewer)